tracker/requirements.txt
gunicorn
Django==3.1.3
djangorestframework==3.12.2

psycopg2

# tests
django_nose==1.4.6
coverage==4.5.4
